requirements.txt
supabase
openai
python-dotenv
langchain
gitpython
mcp
flask